# BasicApi.IntegrationTests/Snapshots/openapi.json
{
  "components": {
    "schemas": {
...
      "AuthResponseDto": {
        "additionalProperties": false,
        "properties": {
          "displayName": {
            "nullable": true,
            "type": "string"
          },
          "email": {
            "nullable": true,
            "type": "string"
          },
          "expiresAt": {
            "description": "When BasicApi.Features.Auth.AuthResponseDto.Token expires (UTC).",
            "format": "date-time",
            "type": "string"
          },
          "refreshToken": {
            "description": "Long-lived opaque token used to obtain a new access token via\n`POST /api/auth/refresh`. Rotated on every refresh — always store the\nlatest value and never send it anywhere except the refresh endpoint.",
            "nullable": true,
            "type": "string"
          },
          "refreshTokenExpiresAt": {
            "description": "When BasicApi.Features.Auth.AuthResponseDto.RefreshToken expires (UTC). After that the user must log in again.",
            "format": "date-time",
            "type": "string"
          },
          "token": {
            "description": "Short-lived JWT for API and hub calls. Send as `Authorization: Bearer`.",
            "nullable": true,
            "type": "string"
          },
          "userId": {
            "format": "uuid",
            "type": "string"
          },
          "username": {
            "nullable": true,
            "type": "string"
          }
        },
        "type": "object"
      },
      "ChangePasswordDto": {
        "additionalProperties": false,
        "properties": {
          "currentPassword": {
            "minLength": 1,
            "type": "string"
          },
          "newPassword": {
            "description": "At least 6 characters and at most 72 bytes in UTF-8, as at registration.",
            "minLength": 6,
            "type": "string"
          }
        },
        "required": [
          "currentPassword",
          "newPassword"
        ],
        "type": "object"
      },
...
      "LoginRequestDto": {
        "additionalProperties": false,
        "properties": {
          "password": {
            "minLength": 6,
            "type": "string"
          },
          "usernameOrEmail": {
            "minLength": 1,
            "type": "string"
          }
        },
        "required": [
          "password",
          "usernameOrEmail"
        ],
        "type": "object"
      },
      "LogoutRequestDto": {
        "additionalProperties": false,
        "properties": {
          "refreshToken": {
            "description": "Refresh token of the session to end. Optional: when omitted, only the\naccess token is discarded client-side and the session stays alive until\nit expires — so clients should always send it.",
            "nullable": true,
            "type": "string"
          }
        },
        "type": "object"
      },
...
      "ProblemDetails": {
        "additionalProperties": {},
        "properties": {
          "detail": {
            "nullable": true,
            "type": "string"
          },
          "instance": {
            "nullable": true,
            "type": "string"
          },
          "status": {
            "format": "int32",
            "nullable": true,
            "type": "integer"
          },
          "title": {
            "nullable": true,
            "type": "string"
          },
          "type": {
            "nullable": true,
            "type": "string"
          }
        },
        "type": "object"
      },
...
      "RefreshTokenRequestDto": {
        "additionalProperties": false,
        "properties": {
          "refreshToken": {
            "minLength": 1,
            "type": "string"
          }
        },
        "required": [
          "refreshToken"
        ],
        "type": "object"
      },
      "RegisterRequestDto": {
        "additionalProperties": false,
        "properties": {
          "displayName": {
            "maxLength": 100,
            "nullable": true,
            "type": "string"
          },
          "email": {
            "format": "email",
            "minLength": 1,
            "type": "string"
          },
          "password": {
            "minLength": 6,
            "type": "string"
          },
          "username": {
            "maxLength": 50,
            "minLength": 3,
            "type": "string"
          }
        },
        "required": [
          "email",
          "password",
          "username"
        ],
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "Bearer": {
        "bearerFormat": "JWT",
        "description": "Enter your JWT token. Example: 'eyJhbGci...'",
        "scheme": "Bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "contact": {
      "email": "support@myapi.com",
      "name": "API Support Team",
      "url": "https://myapi.com"
    },
    "description": "Real-time Chat API with JWT authentication",
    "title": "Chat API",
    "version": "v1"
  },
  "openapi": "3.0.4",
  "paths": {
    "/api/Auth/login": {
      "post": {
        "description": "Returns a short-lived `token` (access) and a long-lived `refreshToken`.\nStore both; when the access token expires, call `POST /api/auth/refresh`\ninstead of asking for the password again.",
        "requestBody": {
          "content": {
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequestDto"
              }
            },
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequestDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequestDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthResponseDto"
                }
              }
            },
            "description": "OK"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Bad Request"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Unauthorized"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Too Many Requests"
          }
        },
        "summary": "Authenticate a user",
        "tags": [
          "Authentication"
        ]
      }
    },
    "/api/Auth/logout": {
      "post": {
        "description": "Send the session's `refreshToken` in the body — that is what actually gets revoked.\nWithout it the server has nothing to invalidate and the session stays alive until\nit expires.\n            \nIdempotent: an unknown or already-revoked token also returns 200, so the endpoint\ncannot be used to find out which tokens exist. The access token keeps working until\nit expires (minutes) — discard it client-side.",
        "requestBody": {
          "content": {
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/LogoutRequestDto"
              }
            },
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LogoutRequestDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/LogoutRequestDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Unauthorized"
          }
        },
        "summary": "Log out of the current session.",
        "tags": [
          "Authentication"
        ]
      }
    },
    "/api/Auth/logout-all": {
      "post": {
        "description": "Revokes all refresh tokens of the current user — use it after a password change or\nwhen a device is lost. Already-issued access tokens stay valid until they expire.",
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Unauthorized"
          }
        },
        "summary": "Log out of every session on all devices.",
        "tags": [
          "Authentication"
        ]
      }
    },
    "/api/Auth/password": {
      "post": {
        "description": "Needs the current password. Every other sign-in of the user ends at once — their refresh\ntokens stop working and their hub connections close; this device stays signed in. Access\ntokens already issued to other devices work until they expire (minutes). Limited like\nsign-in: 5 attempts a minute from one IP.\n            \nErrors: `400 WRONG_PASSWORD`, `400 PASSWORD_TOO_LONG`, `400 VALIDATION_ERROR`\n(new password shorter than 6), `429 RATE_LIMITED`.",
        "requestBody": {
          "content": {
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/ChangePasswordDto"
              }
            },
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChangePasswordDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/ChangePasswordDto"
              }
            }
          }
        },
        "responses": {
          "204": {
            "description": "No Content"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Bad Request"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Unauthorized"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Too Many Requests"
          }
        },
        "summary": "Change the password.",
        "tags": [
          "Authentication"
        ]
      }
    },
    "/api/Auth/refresh": {
      "post": {
        "description": "Call this when the access token has expired (or is about to). The endpoint is\nanonymous on purpose — the expired access token is not required, and not accepted\nas proof of anything.\n            \n**Rotation:** every successful call invalidates the refresh token you sent and\nreturns a new one. Always persist the new value.\n            \n**Parallel refreshes are safe.** A token that was rotated less than 30 seconds ago\nstill works, so two requests racing after a 401 both succeed instead of logging the\nuser out. Replaying a token after that window is treated as theft and revokes every\nsession in that login's chain (`REFRESH_TOKEN_REUSED`).\n            \nError codes: `INVALID_REFRESH_TOKEN`, `REFRESH_TOKEN_EXPIRED`, `SESSION_REVOKED`,\n`REFRESH_TOKEN_REUSED`, `USER_INACTIVE`, `USER_NOT_FOUND` — all as 401.\nAny of them means the same thing for the client: show the login screen.",
        "requestBody": {
          "content": {
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/RefreshTokenRequestDto"
              }
            },
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RefreshTokenRequestDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/RefreshTokenRequestDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthResponseDto"
                }
              }
            },
            "description": "OK"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Bad Request"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Unauthorized"
          }
        },
        "summary": "Exchange a refresh token for a new access/refresh pair.",
        "tags": [
          "Authentication"
        ]
      }
    },
    "/api/Auth/register": {
      "post": {
        "description": "Registration signs the user in: the response carries the same access and\nrefresh tokens as `POST /api/auth/login`.",
        "requestBody": {
          "content": {
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterRequestDto"
              }
            },
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterRequestDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterRequestDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthResponseDto"
                }
              }
            },
            "description": "Created"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Bad Request"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Conflict"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Too Many Requests"
          }
        },
        "summary": "Register a new user",
        "tags": [
          "Authentication"
        ]
      }
    },
...
  },
  "security": [
    {
      "Bearer": []
    }
  ],
...
}